Added ability to specify code blocks with raw formatting.
Fix for issue : https://github.com/RvonMassow/xDoc/issues/94

This check-in excludes a lot of boilerplate classes generated when
running the MWE2 file for ease of review. The only model element 
modified is CodeBlock (adding a boolean field called 'raw'), and the
only three changes are in StatefulEclipseHelpGenerator.xtend,
Xdoc.xtext, XdocFormatter.java and in various auto-generated
lexing/parsing files.

The default behaviour is maintained for code[....] blocks, but now
code-raw[...] is supported, with the user having fine-grained control
over the idents/white-space handling (1 to 1).